refactor: migrate scripts to typscript
