+++ package.json
@@ -1,5 +1,6 @@
 {
   "name": "yt_gaming_library",
+  "type": "module",
   "version": "0.1.0",
   "private": true,
   "scripts": {
@@ -10,7 +11,7 @@
     "generate-responsive-images": "node scripts/generate-responsive-images.mjs",
     "sample-game": "npm run generate-responsive-images -- singleGame PLRfhDHeBTBJ4KQEvirhf9p1o_xuFysva8 covers cover.jpg",
     "generate-playlist-data-report": "node scripts/generate-playlists-stats.mjs",
-    "generate-playlist-csv": "node scripts/generate-playlist-csv.mjs",
+    "generate-playlist-csv": "tsx scripts/generate-playlist-csv.ts",
     "find-published-games": "tsx scripts/findPublishedGames.ts",
     "generate-api-json-files": "tsx scripts/generateJsonFiles.ts"
   },
